+++ platform-operator/controllers/verrazzano/component/clusterapi/clusterapi.go
@@ -5,10 +5,16 @@ package clusterapi
 
 import (
 	"bytes"
-	"os"
-	"text/template"
-
+	"context"
+	"fmt"
+	"github.com/verrazzano/verrazzano/pkg/log/vzlog"
 	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component/spi"
+	v1 "k8s.io/api/core/v1"
+	"os"
+	clusterapi "sigs.k8s.io/cluster-api/cmd/clusterctl/client"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+	"strings"
+	"text/template"
 )
 
 const clusterctlYamlTemplate = `
@@ -45,10 +51,14 @@ providers:
 `
 
 const (
-	clusterTopology         = "CLUSTER_TOPOLOGY"
-	expClusterResourceSet   = "EXP_CLUSTER_RESOURCE_SET"
-	expMachinePool          = "EXP_MACHINE_POOL"
-	initOCIClientsOnStartup = "INIT_OCI_CLIENTS_ON_STARTUP"
+	clusterTopology                           = "CLUSTER_TOPOLOGY"
+	expClusterResourceSet                     = "EXP_CLUSTER_RESOURCE_SET"
+	expMachinePool                            = "EXP_MACHINE_POOL"
+	initOCIClientsOnStartup                   = "INIT_OCI_CLIENTS_ON_STARTUP"
+	clusterAPIControllerImage                 = "cluster-api-controller"
+	clusterAPIOCIControllerImage              = "cluster-api-oci-controller"
+	clusterAPIOCNEBoostrapControllerImage     = "cluster-api-ocne-bootstrap-controller"
+	clusterAPIOCNEControlPLaneControllerImage = "cluster-api-ocne-control-plane-controller"
 )
 
 type ImageConfig struct {
@@ -57,8 +67,25 @@ type ImageConfig struct {
 	Tag        string
 }
 
+// PodMatcherClusterAPI matches pods with an out of date CoreProvider, BootstrapProvider, ControlPlaneProvider, or InfrastructureProvider.
+type PodMatcherClusterAPI struct {
+	coreProvider           string
+	bootstrapProvider      string
+	controlPlaneProvider   string
+	infrastructureProvider string
+}
+
+type ImageCheckResult int
+
+const (
+	OutOfDate ImageCheckResult = iota
+	UpToDate
+	NotFound
+)
+
 const (
 	defaultClusterAPIDir = "/verrazzano/.cluster-api"
+	clusterAPIDirEnv     = "VERRAZZANO_CLUSTER_API_DIR"
 )
 
 var clusterAPIDir = defaultClusterAPIDir
@@ -69,6 +96,13 @@ func setClusterAPIDir(dir string) {
 }
 func resetClusterAPIDir() {
 	clusterAPIDir = defaultClusterAPIDir
+}
+
+func getClusterAPIDir() string {
+	if capiDir := os.Getenv(clusterAPIDirEnv); len(capiDir) > 0 {
+		return capiDir
+	}
+	return clusterAPIDir
 }
 
 // preInstall implementation for the ClusterAPI Component
@@ -131,7 +165,7 @@ func createClusterctlYaml(ctx spi.ComponentContext) error {
 		return err
 	}
 
-	err = os.Mkdir(clusterAPIDir, 0755)
+	err = os.Mkdir(getClusterAPIDir(), 0755)
 	if err != nil {
 		if !os.IsExist(err) {
 			return err
@@ -139,7 +173,7 @@ func createClusterctlYaml(ctx spi.ComponentContext) error {
 	}
 
 	// Create the clusterctl.yaml used when initializing CAPI.
-	return os.WriteFile(clusterAPIDir+"/clusterctl.yaml", result.Bytes(), 0600)
+	return os.WriteFile(getClusterAPIDir()+"/clusterctl.yaml", result.Bytes(), 0600)
 }
 
 // applyTemplate applies the CAPI provider image overrides and versions to the clusterctl.yaml template
@@ -159,4 +193,61 @@ func applyTemplate(templateContent string, params interface{}) (bytes.Buffer, er
 
 	// Return the result containing the processed template
 	return buf, nil
+}
+
+// Initializes Cluster API controller image versions considering overrides from VZ CR
+func (c *PodMatcherClusterAPI) initializeImageVersionsOverrides(log vzlog.VerrazzanoLogger, overrides OverridesInterface) error {
+	c.coreProvider = overrides.GetClusterAPIControllerFullImagePath()
+	c.infrastructureProvider = overrides.GetOCIControllerFullImagePath()
+	c.bootstrapProvider = overrides.GetOCNEBootstrapControllerFullImagePath()
+	c.controlPlaneProvider = overrides.GetOCNEControlPlaneControllerFullImagePath()
+	return nil
+}
+
+// isImageOutOfDate returns true if the container image is not as expected (out of date)
+func isImageOutOfDate(log vzlog.VerrazzanoLogger, imageName, actualImage, expectedImage string) ImageCheckResult {
+	log.Infof("Checking if the image %v is out of date,  actual image: %v, expected image: %v", imageName, actualImage, expectedImage)
+	if strings.Contains(actualImage, imageName) {
+		if 0 != strings.Compare(actualImage, expectedImage) {
+			return OutOfDate
+		}
+		return UpToDate
+	}
+	return NotFound
+}
+
+// MatchAndPrepareUpgradeOptions when a pod has an outdated cluster api controllers images and prepares upgrade options for outdated images.
+func (c *PodMatcherClusterAPI) matchAndPrepareUpgradeOptions(ctx spi.ComponentContext, overrides OverridesInterface) (clusterapi.ApplyUpgradeOptions, error) {
+	c.initializeImageVersionsOverrides(ctx.Log(), overrides)
+	applyUpgradeOptions := clusterapi.ApplyUpgradeOptions{}
+	podList := &v1.PodList{}
+	if err := ctx.Client().List(context.TODO(), podList, &client.ListOptions{Namespace: ComponentNamespace}); err != nil {
+		return applyUpgradeOptions, err
+	}
+	const formatString = "%s/%s:%s"
+	for _, pod := range podList.Items {
+		for _, co := range pod.Spec.Containers {
+			if isImageOutOfDate(ctx.Log(), clusterAPIControllerImage, co.Image, c.coreProvider) == OutOfDate {
+				applyUpgradeOptions.CoreProvider = fmt.Sprintf(formatString, ComponentNamespace, clusterAPIProviderName, overrides.GetClusterAPIVersion())
+			}
+			if isImageOutOfDate(ctx.Log(), clusterAPIOCNEBoostrapControllerImage, co.Image, c.bootstrapProvider) == OutOfDate {
+				applyUpgradeOptions.BootstrapProviders = append(applyUpgradeOptions.BootstrapProviders, fmt.Sprintf(formatString, ComponentNamespace, ocneProviderName, overrides.GetOCNEBootstrapVersion()))
+			}
+			if isImageOutOfDate(ctx.Log(), clusterAPIOCNEControlPLaneControllerImage, co.Image, c.controlPlaneProvider) == OutOfDate {
+				applyUpgradeOptions.ControlPlaneProviders = append(applyUpgradeOptions.ControlPlaneProviders, fmt.Sprintf(formatString, ComponentNamespace, ocneProviderName, overrides.GetOCNEControlPlaneVersion()))
+			}
+			if isImageOutOfDate(ctx.Log(), clusterAPIOCIControllerImage, co.Image, c.infrastructureProvider) == OutOfDate {
+				applyUpgradeOptions.InfrastructureProviders = append(applyUpgradeOptions.InfrastructureProviders, fmt.Sprintf(formatString, ComponentNamespace, ociProviderName, overrides.GetOCIVersion()))
+			}
+		}
+	}
+	return applyUpgradeOptions, nil
+}
+
+// isUpgradeOptionsNotEmpty returns true if any of the options is not empty
+func isUpgradeOptionsNotEmpty(upgradeOptions clusterapi.ApplyUpgradeOptions) bool {
+	return len(upgradeOptions.CoreProvider) != 0 ||
+		len(upgradeOptions.BootstrapProviders) != 0 ||
+		len(upgradeOptions.ControlPlaneProviders) != 0 ||
+		len(upgradeOptions.InfrastructureProviders) != 0
 }