+++ platform-operator/controllers/verrazzano/component/clusterapi/clusterapi_overrides.go
@@ -5,16 +5,15 @@ package clusterapi
 
 import (
 	"fmt"
-	"os"
-	"path/filepath"
-	"strings"
-
 	"github.com/verrazzano/verrazzano/pkg/bom"
 	vzyaml "github.com/verrazzano/verrazzano/pkg/yaml"
 	"github.com/verrazzano/verrazzano/platform-operator/constants"
 	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component/common/override"
 	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component/spi"
 	"github.com/verrazzano/verrazzano/platform-operator/internal/config"
+	"os"
+	"path"
+	"path/filepath"
 	"sigs.k8s.io/yaml"
 )
 
@@ -36,11 +35,11 @@ type defaultProviders struct {
 }
 
 type capiProvider struct {
-	Image         capiImage `json:"image,omitempty"`
-	Version       string    `json:"version,omitempty"`
-	URL           string    `json:"url,omitempty"`
-	Name          string    `json:"name,omitempty"`
-	MetaddataFile string    `json:"metaddataFile,omitempty"`
+	Image        capiImage `json:"image,omitempty"`
+	Version      string    `json:"version,omitempty"`
+	URL          string    `json:"url,omitempty"`
+	Name         string    `json:"name,omitempty"`
+	MetadataFile string    `json:"metadataFile,omitempty"`
 }
 
 type capiImage struct {
@@ -54,18 +53,22 @@ type capiImage struct {
 type OverridesInterface interface {
 	GetGlobalRegistry() string
 	GetClusterAPIRepository() string
+	GetClusterAPIControllerFullImagePath() string
 	GetClusterAPITag() string
 	GetClusterAPIURL() string
 	GetClusterAPIVersion() string
 	GetOCIRepository() string
+	GetOCIControllerFullImagePath() string
 	GetOCITag() string
 	GetOCIURL() string
 	GetOCIVersion() string
 	GetOCNEBootstrapRepository() string
+	GetOCNEBootstrapControllerFullImagePath() string
 	GetOCNEBootstrapTag() string
 	GetOCNEBootstrapURL() string
 	GetOCNEBootstrapVersion() string
 	GetOCNEControlPlaneRepository() string
+	GetOCNEControlPlaneControllerFullImagePath() string
 	GetOCNEControlPlaneTag() string
 	GetOCNEControlPlaneURL() string
 	GetOCNEControlPlaneVersion() string
@@ -143,6 +146,22 @@ func (c capiOverrides) GetOCNEControlPlaneVersion() string {
 	return getProviderVersion(c.DefaultProviders.OCNEControlPlane)
 }
 
+func (c capiOverrides) GetClusterAPIControllerFullImagePath() string {
+	return fmt.Sprintf("%s/%s:%s", c.GetClusterAPIRepository(), clusterAPIControllerImage, c.GetClusterAPITag())
+}
+
+func (c capiOverrides) GetOCIControllerFullImagePath() string {
+	return fmt.Sprintf("%s/%s:%s", c.GetOCIRepository(), clusterAPIOCIControllerImage, c.GetOCITag())
+}
+
+func (c capiOverrides) GetOCNEBootstrapControllerFullImagePath() string {
+	return fmt.Sprintf("%s/%s:%s", c.GetOCNEBootstrapRepository(), clusterAPIOCNEBoostrapControllerImage, c.GetOCNEBootstrapTag())
+}
+
+func (c capiOverrides) GetOCNEControlPlaneControllerFullImagePath() string {
+	return fmt.Sprintf("%s/%s:%s", c.GetOCNEControlPlaneRepository(), clusterAPIOCNEControlPLaneControllerImage, c.GetOCNEControlPlaneTag())
+}
+
 // getRepositoryForProvider - return the repository in the format that clusterctl
 // expects (registry/owner)
 func getRepositoryForProvider(overrides capiOverrides, provider capiProvider) string {
@@ -176,10 +195,10 @@ func getURLForProvider(provider capiProvider, ownerRepo string) string {
 		return provider.URL
 	}
 	if len(provider.Version) > 0 {
-		return formatProviderURL(true, provider.Image.Repository, ownerRepo, provider.Version, provider.MetaddataFile)
+		return formatProviderURL(true, provider.Image.Repository, ownerRepo, provider.Version, provider.MetadataFile)
 	}
 	// Return default value
-	return formatProviderURL(false, provider.Image.Repository, provider.Name, provider.Image.BomVersion, provider.MetaddataFile)
+	return formatProviderURL(false, provider.Image.Repository, provider.Name, provider.Image.BomVersion, provider.MetadataFile)
 }
 
 // formatProviderURL - return the provider URL using the following format
@@ -188,7 +207,12 @@ func formatProviderURL(remote bool, owner string, repo string, version string, m
 	if remote {
 		return fmt.Sprintf("https://github.com/%s/%s/releases/%s/%s", owner, repo, version, metadataFile)
 	}
-	return fmt.Sprintf("/verrazzano/capi/%s/%s/%s", repo, version, metadataFile)
+
+	var capiRoot = "/verrazzano/capi"
+	if _, err := os.Stat(capiRoot); err != nil {
+		capiRoot = path.Join(config.GetPlatformDir(), "capi")
+	}
+	return fmt.Sprintf("%s/%s/%s/%s", capiRoot, repo, version, metadataFile)
 }
 
 // createOverrides - create the overrides input for install/upgrade of the
@@ -228,13 +252,13 @@ func getBaseOverrides() (*capiOverrides, error) {
 
 	// Initialize internal static values
 	overrides.DefaultProviders.Core.Name = "cluster-api"
-	overrides.DefaultProviders.Core.MetaddataFile = "core-components.yaml"
+	overrides.DefaultProviders.Core.MetadataFile = "core-components.yaml"
 	overrides.DefaultProviders.OCI.Name = "infrastructure-oci"
-	overrides.DefaultProviders.OCI.MetaddataFile = "infrastructure-components.yaml"
+	overrides.DefaultProviders.OCI.MetadataFile = "infrastructure-components.yaml"
 	overrides.DefaultProviders.OCNEBootstrap.Name = "bootstrap-ocne"
-	overrides.DefaultProviders.OCNEBootstrap.MetaddataFile = "bootstrap-components.yaml"
+	overrides.DefaultProviders.OCNEBootstrap.MetadataFile = "bootstrap-components.yaml"
 	overrides.DefaultProviders.OCNEControlPlane.Name = "control-plane-ocne"
-	overrides.DefaultProviders.OCNEControlPlane.MetaddataFile = "control-plane-components.yaml"
+	overrides.DefaultProviders.OCNEControlPlane.MetadataFile = "control-plane-components.yaml"
 
 	return overrides, err
 }
@@ -254,7 +278,7 @@ func mergeBOMOverrides(ctx spi.ComponentContext, overrides *capiOverrides) error
 
 	// Populate core provider values
 	core := &overrides.DefaultProviders.Core.Image
-	imageConfig, err := getImageOverride(ctx, bomFile, "capi-cluster-api", "")
+	imageConfig, err := getImageOverride(ctx, bomFile, "capi-cluster-api", "capi-cluster-api", "cluster-api-controller")
 	if err != nil {
 		return err
 	}
@@ -262,7 +286,7 @@ func mergeBOMOverrides(ctx spi.ComponentContext, overrides *capiOverrides) error
 
 	// Populate OCI provider values
 	oci := &overrides.DefaultProviders.OCI.Image
-	imageConfig, err = getImageOverride(ctx, bomFile, "capi-oci", "")
+	imageConfig, err = getImageOverride(ctx, bomFile, "capi-oci", "capi-oci", "cluster-api-oci-controller")
 	if err != nil {
 		return err
 	}
@@ -270,7 +294,7 @@ func mergeBOMOverrides(ctx spi.ComponentContext, overrides *capiOverrides) error
 
 	// Populate bootstrap provider values
 	bootstrap := &overrides.DefaultProviders.OCNEBootstrap.Image
-	imageConfig, err = getImageOverride(ctx, bomFile, "capi-ocne", "cluster-api-ocne-bootstrap-controller")
+	imageConfig, err = getImageOverride(ctx, bomFile, "capi-ocne", "capi-ocne", "cluster-api-ocne-bootstrap-controller")
 	if err != nil {
 		return err
 	}
@@ -278,7 +302,7 @@ func mergeBOMOverrides(ctx spi.ComponentContext, overrides *capiOverrides) error
 
 	// Populate controlPlane provider values
 	controlPlane := &overrides.DefaultProviders.OCNEControlPlane.Image
-	imageConfig, err = getImageOverride(ctx, bomFile, "capi-ocne", "cluster-api-ocne-control-plane-controller")
+	imageConfig, err = getImageOverride(ctx, bomFile, "capi-ocne", "capi-ocne", "cluster-api-ocne-control-plane-controller")
 	if err != nil {
 		return err
 	}
@@ -331,35 +355,25 @@ func mergeUserOverrides(ctx spi.ComponentContext, overrides *capiOverrides) erro
 }
 
 // getImageOverride returns the image override and version for a given CAPI provider.
-func getImageOverride(ctx spi.ComponentContext, bomFile bom.Bom, component string, imageName string) (image *ImageConfig, err error) {
+func getImageOverride(ctx spi.ComponentContext, bomFile bom.Bom, component string, subcomponent string, imageName string) (image *ImageConfig, err error) {
 	version, err := bomFile.GetComponentVersion(component)
 	if err != nil {
 		return nil, err
 	}
 
-	images, err := bomFile.GetImageNameList(component)
+	subComp, err := bomFile.GetSubcomponent(subcomponent)
 	if err != nil {
 		return nil, err
 	}
 
-	var repository string
-	var tag string
-	for _, image := range images {
-		if len(imageName) == 0 || strings.Contains(image, imageName) {
-			imageSplit := strings.Split(image, ":")
-			tag = imageSplit[1]
-			index := strings.LastIndex(imageSplit[0], "/")
-			repository = imageSplit[0][:index]
-			repoSplit := strings.Split(repository, "/")
-			repository = strings.TrimPrefix(repository, repoSplit[0])
-			repository = strings.TrimPrefix(repository, "/")
-			break
-		}
+	img, err := bomFile.FindImage(subComp, imageName)
+	if err != nil {
+		return nil, err
 	}
 
-	if len(repository) == 0 || len(tag) == 0 {
+	repository := bomFile.ResolveRepo(subComp, img)
+	if len(repository) == 0 || len(img.ImageTag) == 0 {
 		return nil, ctx.Log().ErrorNewErr("Failed to find image override for %s/%s", component, imageName)
 	}
-
-	return &ImageConfig{Version: version, Repository: repository, Tag: tag}, nil
+	return &ImageConfig{Version: version, Repository: repository, Tag: img.ImageTag}, nil
 }